Gracefully handle when an alias doesn't exist
If an alias wasn't found the script would fall over. Now it handles it
gracefully, passing over it and continuing.

+++ scripts/update-index-alias.py
@@ -8,11 +8,11 @@ access the credentials repo.
 To delete the old index (the index losing the '<alias>-old' alias), set --delete-old-index=yes
 
 Usage:
-scripts/update-index-alias.py <alias> <target> <search-api-endpoint> [--stage=<stage> --delete-old-index=<value>]
+scripts/update-index-alias.py <alias> <target> <search-api-endpoint> [options]
 
 Options:
-    --stage=<stage> [default: development]
-    --delete-old-index=<delete-old-index>  [default: no]
+    --stage=stage               The stage being updated [default: development]
+    --delete-old-index=yes/no   Whether to delete the index losing its alias [default: no]
 """
 import sys
 import json
@@ -35,10 +35,13 @@ def update_index_alias(alias, target, stage, endpoint, delete_old_index):
 
     current_aliased_index = _get_index_from_alias(alias, endpoint)
     old_aliased_index = _get_index_from_alias(alias_old, endpoint)
-    _apply_alias_to_index(alias, target, endpoint, headers)
-    _apply_alias_to_index(alias_old, current_aliased_index, endpoint, headers)
 
-    if delete_old_index:
+    _apply_alias_to_index(alias, target, endpoint, headers)
+
+    if current_aliased_index:
+        _apply_alias_to_index(alias_old, current_aliased_index, endpoint, headers)
+
+    if old_aliased_index and delete_old_index:
         _delete_index(old_aliased_index, endpoint, auth_token)
 
 
@@ -49,9 +52,13 @@ def _get_index_from_alias(alias, endpoint):
     _check_response_status(response, 'fetching indexes')
 
     all_indexes = json.loads(response.content)['es_status']
-    index_name = [index for index in all_indexes.keys() if alias in all_indexes[index]['aliases']][0]
+    index_name = [index for index in all_indexes.keys() if alias in all_indexes[index]['aliases']]
 
-    return index_name
+    if not index_name:
+        print("No index found with {} alias".format(alias))
+        return None
+
+    return index_name[0]
 
 
 def _apply_alias_to_index(alias, target, endpoint, headers):
